+++ js/src/admin/index.tsx
@@ -1,8 +1,8 @@
 import app from 'flarum/admin/app';
-import RecycleBinPage from './components/RecycleBinPage';
+import SettingsPage from './components/SettingsPage';
 
 app.initializers.add('walsgit/recycle-bin', () => {
         app.extensionData
             .for('walsgit-recycle-bin')
-            .registerPage(RecycleBinPage as any);
+            .registerPage(SettingsPage as any);
 });